+++ Vanny/app/src/main/java/com/example/vanny/VideoPage.java
@@ -5,12 +5,14 @@ import androidx.appcompat.app.AppCompatActivity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
+import android.widget.Button;
 import android.widget.ImageButton;
 
 
 public class VideoPage extends AppCompatActivity {
 
     private ImageButton backBtn;
+    private Button reportFileBtn;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -23,10 +25,23 @@ public class VideoPage extends AppCompatActivity {
                 openActivity2();
             }
         });
+
+        reportFileBtn = (Button) findViewById(R.id.videoOne);
+        reportFileBtn.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                openActivity3();
+            }
+        });
     }
 
     public void openActivity2(){
         Intent intent = new Intent(this, SettingsPage.class);
+        startActivity(intent);
+    }
+
+    public void openActivity3(){
+        Intent intent = new Intent(this, ReportFile.class);
         startActivity(intent);
     }
 }
